Since the last commit I'm not entirely sure what's been done because I let a few commits build up over a week or so (while I did very little - it was around the time of handing in the proposal).
Basically, AStar has been implemented (and hence changes to class Node
etc to add g and f values), and a Test class (within SRC) has been
added. I may add a new implementation that is closer to the pseudocode
for A* provided with the explanation of Theta* so that the Theta* code
is easier to understand in comparison!